refactor(external/dependency) : 외부 의존성 인터페이스 or 유틸로 격리
