+++ src/main/java/baseball/Application.java
@@ -1,5 +1,8 @@
 package baseball;
 
+import baseball.factory.ComputerFactory;
+import baseball.factory.GameRestartPolicyFactory;
+import baseball.factory.PlayerFactory;
 import baseball.game.Computer;
 import baseball.game.Game;
 import baseball.game.GameRestartPolicy;
@@ -9,9 +12,9 @@ import baseball.game.Player;
 public class Application {
     public static void main(String[] args) {
         Game game = new Game(
-            Computer.create(),
-            Player.create(),
-            GameRestartPolicy.create()
+            ComputerFactory.create(),
+            PlayerFactory.create(),
+            GameRestartPolicyFactory.create()
         );
         new GameStarter(game).start();
     }
