+++ src/main/java/baseball/factory/PlayerFactory.java
@@ -1,0 +1,15 @@
+package baseball.factory;
+
+import baseball.game.Player;
+import baseball.io.ConsolePrinter;
+import baseball.io.ConsoleScanner;
+
+public class PlayerFactory {
+
+    public static Player create() {
+        return new Player(
+            new ConsolePrinter(),
+            new ConsoleScanner()
+        );
+    }
+}
