+++ src/main/java/baseball/game/GameRestartPolicy.java
@@ -2,15 +2,16 @@ package baseball.game;
 
 import baseball.io.Printer;
 import baseball.context.RestartPolicy;
-import baseball.io.ConsolePrinter;
-import camp.nextstep.edu.missionutils.Console;
+import baseball.io.Scanner;
 
 public class GameRestartPolicy implements RestartPolicy {
 
     private final Printer printer;
+    private final Scanner scanner;
 
-    public GameRestartPolicy(Printer printer) {
+    public GameRestartPolicy(Printer printer, Scanner scanner) {
         this.printer = printer;
+        this.scanner = scanner;
     }
 
     @Override
@@ -18,7 +19,7 @@ public class GameRestartPolicy implements RestartPolicy {
         if (canRestart) {
             printer.println("3개의 숫자를 모두 맞히셨습니다! 게임 종료");
             printer.println("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
-            String signal = Console.readLine();
+            String signal = scanner.readLine();
             onSignal("1".equals(signal), restartSignal);
             onSignal("2".equals(signal), exitSignal);
         }
@@ -28,11 +29,5 @@ public class GameRestartPolicy implements RestartPolicy {
         if (on) {
             signal.run();
         }
-    }
-
-    public static GameRestartPolicy create() {
-        return new GameRestartPolicy(
-            new ConsolePrinter()
-        );
     }
 }
