+++ src/main/java/baseball/game/Player.java
@@ -2,16 +2,17 @@ package baseball.game;
 
 import baseball.context.Numbers;
 import baseball.io.Printer;
-import baseball.io.ConsolePrinter;
+import baseball.io.Scanner;
 import baseball.util.NumbersValidator;
-import camp.nextstep.edu.missionutils.Console;
 
 public class Player {
 
     private final Printer printer;
+    private final Scanner scanner;
 
-    public Player(Printer printer) {
+    public Player(Printer printer, Scanner scanner) {
         this.printer = printer;
+        this.scanner = scanner;
     }
 
     public Numbers present() {
@@ -20,14 +21,8 @@ public class Player {
 
     private Numbers printAndGetNumbers() {
         printer.print("숫자를 입력해주세요 : ");
-        char[] rawNumbers = Console.readLine().toCharArray();
+        char[] rawNumbers = scanner.readLine().toCharArray();
         NumbersValidator.validate(rawNumbers);
         return Numbers.valueOf(rawNumbers);
-    }
-
-    public static Player create() {
-        return new Player(
-            new ConsolePrinter()
-        );
     }
 }
