+++ src/main/java/baseball/util/RandomUtil.java
@@ -1,0 +1,10 @@
+package baseball.util;
+
+import camp.nextstep.edu.missionutils.Randoms;
+
+public class RandomUtil {
+
+    public static int getNumber() {
+        return Randoms.pickNumberInRange(1, 9);
+    }
+}
